+++ changelog.md
@@ -3,6 +3,8 @@
 ## Version 0.54.0 (2023-07-08)
 ### New Features
 1. Import deck via URL
+### Fixes
+1. Fix "Collect All" frienship reward button
 
 ## Version 0.53.1 (2023-07-02)
 1. Fixed changelog